Change call stack searching code [revdep skip]
lapply and sys.calls dont play well with other in devtools::test and other testthat variants
that are used to test the package on travis. It works for individual test files. However, to
avoid the issue, the lapply call has been changed to sapply. There is a note in the help file
about the issues with lapply