DS 3136 SumRows and SumColumns (#5)
* DS-3136 First draft of Sum

* DS-3136 Simplify and improve previous version

* DS-3136 Add first draft of SumRows

* DS-3136 Add first draft of SumColumns

* DS-3136 Add tests for utility functions

* DS-3136 Update warning message to be more flexible

* DS-3136 Update sum row and column tests

* DS-3136 Improve efficiency of opp Inf check

* DS-3136 Add feature to match names

* DS-3136 Minor documentation update

* DS-3136 Update function.name handling

* DS-3136 Refactor checking multiple statistics

* DS-3136 Refactor SumRows

* DS-3136 Modify SumColumns for new behaviour

* DS-3136 Simplify extraction of chart data

* DS-3136 Remove old Inf check function

* DS-3136 Update element matching functions [ci skip]

* DS-3136 Add unit tests for element matching

* DS-3136 Update tests for allowed inputs

* DS-3136 Minor doc update

* DS-3136 Add utils to NAMESPACE

* DS-3136 Add missing argument for function.name

* DS-3136 Remove check of weights in subset code

* DS-3136 Create common Rd file for Sum and SumRows

* DS-3136 Improve tests for Sum

* DS-3136 Placeholder for detailed documentation

* Fix infinity checking and row/col removal

* DS-3136 Update doc for utilities

* DS-3136 Update tests

* fixup! DS-3136 Update tests

* DS-3136 Remove duplicated quoting

* DS-3136 Add sQuote to tests

Use the getFancyQuote option of the user and use sQuote in tests

* DS-3136 Add basic SumRows fuzzy tests

* DS-3136-2 Levenshtein improvements

* DS-3136 Fix mapping error in unmatched fuzzy

* DS-3136 Error early if unmatched observed and cannot be ignored

* DS-3136 Improve SumColumns tests

* DS-3136 Handle fuzzy matching with unmatched out

* DS-3136 Fix unit tests

* DS-3136 Add Rd files of main help file

* DS-3136-2 Consolidate Opposite Inf checks for Sum

* DS-3136 Improve warnings for SumRows

* DS-3136-2 Generalize the check for Opposite Infinities

* Update .travis.yml

* Simplify Opp Infinity check

* DS-3136 Rename subset and weighting function

* DS-3136 Update documentation

* Move test data to testthat dir

* Add logging to travis.yml

* Fix spacing last commit/travis cfg edit

* Add aliases for debugging travis [untracked]

* Remove spaces in alias cmd def in travis.yml

* Restore travis.yml [rev dep skip]

* Add install logging to travis [rev dep skip]

* Fix typos in documentationa [ci skip]

* Start fresh

* DS-3136-3 Update Sum signature

Sum should not remove rows and columns as the main use case is element wise.
Also matching will be introduced with the default being an exact match.

* Create simple exact name matching function

Create new function called exactMatchDimensionNames that takes a list of
two character strings and matches them. The output is a list of two named
integer vectors containing the indices to match the elements. Any unmatched
are given the value of NA.

* Improve exact matching tests

* Handle NaN and missing data in Sum

Sum now refactored to do elementwise addition for multiple inputs.
Missing values are now converted to zero if they exist and removal
is requested. Also NaN checks are done elementwise as well.

* Add feature to reshape scalars

* Tidy up getDim function

* Improve reshaping

* Add Reduce Sum template and add tests

* Add unmatched handling for exact matching

* Add tests for fuzzy matching [ci skip]

* Add feature to update dimension labels

If no matching occurs on a dimension, the dimension labels are inspected
and if dimension labels exist for both elements, they are concatenated
with a + separator to denote the dimension with different names has been
added together

* Remove SumRows and SumColumns for now

* DS-3136-3 Update Sum examples

* DS-3136-3 Improve warning and error message

* Fix some tests [ci skip]

* DS-3136 Add code to check call stack

Sum will be used in SumRows and possibly SumColumns. To ensure
an appropriate error or warning is thrown that references the
parent function that called Sum. The code that determines the
function.name variable has been updated

* DS-3136 Add code to remove character statistics

Some statistics are characters such as Columns compared. This
causes the ensure table to take on a character matrix. This
commit checks for character statistics, removes them and then
converts the matrix to numeric via storage.mode

* DS-3136 Bug fixes and test improvements

Found bugs in the existing code and removed them and added unit tests

* DS-3136 Point to dev branches of flipFormat/Transformations

* DS-3136 Add test file for character stat tests

* fixup! DS-3136 Point to dev branches of flipFormat/Transformations

* DS-3136 Restore tests to travis [revdep skip]

* DS-3136 Server independent tests [revdep skip]

Ensure the unit tests pass on both Displayr R server and locally

* DS-3136 Use Sum to compute SumRows [ci skip]

* Split multiple inputs into vectors and add column names [revdep skip]

* For multiple inputs in SumRows, they are split into vectors if possible.
* Where possible, attempt to coerce the output to a column vector with the
names of the multiple inputs as the column name, separated by a +. E.g.
Q1 + Q2 if two inputs had the names Q1 and Q2.

* Update documentation for SumRows [ci skip]

* Tweaks to the output names [revdep skip]

Only add names to some of the split outputs to ensure results render
in a better way

* Change call stack searching code [revdep skip]

lapply and sys.calls dont play well with other in devtools::test and other testthat variants
that are used to test the package on travis. It works for individual test files. However, to
avoid the issue, the lapply call has been changed to sapply. There is a note in the help file
about the issues with lapply

* Add SumColumns function [revdep skip]

Co-authored-by: Mathew W. McLean <mathew.w.mclean@gmail.com>