DS-3136 Add arg to switch off call stack inspection in Sum* (#7) [revdep skip]
* DS-3136 Add support for higher dim Q tables in SumRows/Cols [revdep skip]

* DS-3136 Fix bug where function.name not passed properly in Sum

* DS-3136 Version bump [revdep skip]

* DS-3136 4D+ tables SumRows and SumCols

* DS-3136 Remove obsolete dev branch refs in .travis.yml

* DS-3136 Remove old, not required, test rda files

* DS-3136 Improve regex to handle colons in function call

The Sum suite of functions can be called using verbs:: or verbs:::, the regex
has been improved to capture this situation

* Update travis.yml [revdep skip]

* DS-3136 Improve handling of call stack

The code handling the use of Sum in the call stack was not robust against
non-direct use. E.g. if used via do.call or within lapply. This commit
should identify this and give a sensible output and avoid unintended
side effects.

* fixup! DS-3136 Improve handling of call stack

Relax the unit test to not expect one of testthat functions.
Depending on whether the test is conducted on travis, or executed
via devtools::test or devtools::test_file, the identified function
for the output message will be different

* DS-3136 Add handling of NULL elements

* DS-3136 Tidy up attributes and multiple input checks

* DS-3136 Only subset if necessary

* DS-3136 Address code review comments

* DS-3136 Naming single col outputs in SumColumns

Only determine appropriate name for single column output in SumColumns
if the input is a Variable or Q Table

* DS-3136 Add call. argument to Sum [revdep skip]

* DS-3136 Add call. param to SumRows and SumColumns [revdep skip]

* DS-3136 Update docs [revdep skip]

* DS-3136 Update unit test to work on Displayr servers [revdep skip]

* DS-3136 Fix bug when SumRows called when package not laoded [revdep skip]

Co-authored-by: jrwishart <justin.wishart@idisplayr.com>